Move all but abs_path transforamtion function to seperate script
